refactor: rename and reorganize
