+++ packages/eip712/src/payload/common.ts
@@ -2,7 +2,7 @@ export interface JSONObject {
   [key: string]: any
 }
 
-export interface MessageParams {
+export interface FlattenPayloadResponse {
   payload: JSONObject
   numMessages: number
 }
